MC_dropout/requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
Pillow>=9.5.0
tifffile>=2023.4.12
matplotlib>=3.7.0
tqdm>=4.65.0
pandas>=2.0.0
